repurposed requestbuilder into urlbuilder, more framework stuff on figuring out what needs to be done
