+++ DotaApiCore/DotaAPILib.cs
@@ -10,7 +10,8 @@ namespace DotaApiCore
      */
     class DotaAPIRequest
     {
-        private readonly string _apiKey;
+        private string _apiKey;
+        private URLBuilder httpRequestHelper;
 
         //TODO: custom exceptions?
         public DotaAPIRequest(string APIKey = null)
@@ -19,6 +20,19 @@ namespace DotaApiCore
                 throw new Exception();
             else
                _apiKey = APIKey;
+
+            httpRequestHelper = new URLBuilder();
         }
+
+        /*
+         * Uses contents of this object to send an HTTP GET request from the Dota2 API.
+         * Returns the JSON result upon a successful request, or an error if it fails.
+        */
+        //TODO: fill this in
+        public string SendGetRequest()
+        {
+            return null;
+        }
+
     }
 }
